Initial Commit. See notes for more details
I needed a way to make Unity's GUI system work from a hand controller in VR, for Glycon and a few other things. I figured it out, but it was not easy! So I've simplified and documented my code, and open sourced it. Hopefully this will help other Unity developers, too. This should work across all VR systems. 

Basically, this means you can build your GUI in uGUI, and test it with a mouse. Then when you're ready to go VR, just drop these two files in and presto-blammo.